init: allow expanding the program path in a service declaration
The path of the program to run on a service declaration was assumed to
be always given explicitly and never expanded.

Allowing property expansion in the path is useful for the
init_dev_config service, as the service can be defined as part of the
generic init configuration in system, but the actual binary to run
optionally declared by the vendor as a ro sysprop. If the property
is not set, the service start fails and init goes on.

Using an existing program like `runcon' or `sh' to launch the
program (i.e. to have the sysprop expanded as an argument and not as
the path of the program to run) is not possible in early-init before
apexd bootstrap due to the limitations of the bootstrap bionic.

Bug: 378043923
Bug: 417632996
Test: cvd create & ensure com.google.cf.disabled is not activated

Change-Id: Ief439778e5c8983be4d6d264490aca974b6c5dd7